Updated bounds to allow doubles.

+++ src/main/java/hyness/stl/grammar/STLAbstractSyntaxTreeExtractor.java
@@ -35,7 +35,7 @@ public class STLAbstractSyntaxTreeExtractor extends STLBaseVisitor<TreeNode>{
     public TreeNode visitFormula(STLParser.FormulaContext ctx){
         Operation op = Operation.getCode(ctx.op.getText());
         TreeNode ret = null;
-        int low = -1, high = -1;
+        double low = -1, high = -1;
         switch(op) {
             case OR:
                 ret = new DisjunctionNode(visit(ctx.left), visit(ctx.right));
@@ -50,18 +50,18 @@ public class STLAbstractSyntaxTreeExtractor extends STLBaseVisitor<TreeNode>{
                 ret = new NotNode(visit(ctx.child));
                 break;
             case UNTIL:
-                low = Integer.valueOf(ctx.low.getText());
-                high = Integer.valueOf(ctx.high.getText());
+                low = Double.valueOf(ctx.low.getText());
+                high = Double.valueOf(ctx.high.getText());
                 ret = new UntilNode(visit(ctx.left), visit(ctx.right), low, high);
                 break;
             case EVENT:
-                low = Integer.valueOf(ctx.low.getText());
-                high = Integer.valueOf(ctx.high.getText());
+                low = Double.valueOf(ctx.low.getText());
+                high = Double.valueOf(ctx.high.getText());
                 ret = new EventNode(visit(ctx.child), low, high);
                 break;
             case ALWAYS:
-                low = Integer.valueOf(ctx.low.getText());
-                high = Integer.valueOf(ctx.high.getText());
+                low = Double.valueOf(ctx.low.getText());
+                high = Double.valueOf(ctx.high.getText());
                 ret = new AlwaysNode(visit(ctx.child), low, high);
                 break;
             case CONCAT:
